Have made osg::Transform more extensible via additions of new getLocalToWorldMatrix() and getWorldToLocalMatrix(), computeLocalToWorld() and computeWorldToLocal() methods.
Have updated the CullVisitor, IntersectVisitor and Optimizer to use the
new osg::Transform::getLocalToWorldMatrix() which has the same functionality
as the old getMatrix() but is now supports subclasses of osg::Transform
transparently.

Have added osg::PositionAttitudeTransform as subclass of osg::Transform
which manages the transform as position and attitude via a Vec3 and Quat
respectively.


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@322 16af8721-9629-0410-8352-f15c8da7e697