Fixed a bug in the drive manipulator which was due to it using the local coords of the intersection point with the scene rather than the world coord value for that point.  The code now uses osgUtil::Hit::getWorldIntersectionPoint() method for getting the world coords.
Added support for getWorldIntersectionPoint to Viewer.cpp.

Put and #ifdef around the setting of the default display list visitor so
it nolonger sets in under IRIX. This is a world around to the IR graphics
imbending lighting info into display lists, if the display lists are
created before state is set up it produces lighting artifacts such as
flickering.  Remove the the default display list init visitor removes
these problems, display lists are then built on the fly and drawables
a drawn for the first time.


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@377 16af8721-9629-0410-8352-f15c8da7e697