Added Neil Salter's osgUtx library into the distribution, have put all the orignal seperate headers and source into include/osg/UnitTestFrameWork and src/osg/UnitTestFramework.cpp respectively.  I have done this to keep the include and source directories focused on the scene graph rather than the test framework.
Have added a very simple unit test to Vec3.cpp, which are optionally
compiled in with the OSG_COMPILE_UNIT_TESTS define, which is only
currently defined in debug builds under unix.  It should be fine to
add it in under Windows project files as well.

Finally, a new demo app osgunittests runs the tests and reports success
or failure of the tests.


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@1096 16af8721-9629-0410-8352-f15c8da7e697