From Ruben, added handling of vertex indices.
Note from Robert.  New codes produces errors when running
osgfxbrowser cow.osg so I have elected to keep the original
implementation of compute_basis_vectors around and as the default
compile to keep the CVS version working.  Ruben's new version
of compute_basis_vectors has to be explicitly compiled in by
toggling an #if 1 block to #if 0.


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@3882 16af8721-9629-0410-8352-f15c8da7e697