From Gordon Tomlinson: " This fixes some crashing issues I was having with certain FLT files that only had partial colour palettes
Which are typically found in older 14.x files and file converted through Polytrans or Deep exploration etc.

The code that grabs the color entries in ConvertFromFLT::visitColorPalette was assuming that there were full
palettes always coming in and stepping was out of bounds in certain cases (not all) and thus crashing with
access  violations etc,

Although the normal from Creator is to have 1024 or 512 entries , this is not a  really requirement of the format,  just the way Creator creates its default palettes etc.

Code changed to look at the number of entries and use just those and fills in any missing entries with a default white colour"


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@4456 16af8721-9629-0410-8352-f15c8da7e697