From Wang Lam,"The changes allow users of OpenSceneGraph to call a new function readFontStream() to load fonts from a std::istream, rather than from the local filesystem by name.  Such a call may be used, for example, if the user has a font fetched over a network, or a font available in memory without a correspondng filename.
The changes implement the new function by following the corresponding code
for readFontFile().  readFontStream() reads a stream into memory, and
holds that memory for FreeType.

As a basic test, I mangled the osgtext example to use
readFontStream(std::ifstream("font")) in lieu of a readFontFile call, and
the modified example ran completely."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@4686 16af8721-9629-0410-8352-f15c8da7e697