From Brede Johansen,
"The attached file corrects two small typos in MultiSwitch.cpp.  The
local "values" reference and the "_values" member attribute are of
different types but both are vectors so the size() operator happily
compiles.  A renaming of _values to _masks or similar may be a more
future proof solution but the submission only removes the underscore
in two places.

The switch_1701.flt model part of the Creator gallery revealed this bug."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@5231 16af8721-9629-0410-8352-f15c8da7e697