From Eric Sokolowsky,
"Attached is some fixes for Image::readImageFromCurrentTexture. It was
failing when used with a new Image object (some of the fields were
not set before they were used, and some were set to the wrong values).
A new optional parameter was added to give the desired data type, which
defaults to GL_UNSIGNED_BYTE, so programs should only need a recompile
to work properly."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@5265 16af8721-9629-0410-8352-f15c8da7e697