From Brede Johansen,
"I found some artifacts on some models and traced it back to the
external palette override feature.  There are some changes in the
layout of the external record the loader have to take into
consideration.

The other fix is to remove the OF pools from the user data of the
external ProxyNode when it is no longer needed."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@5281 16af8721-9629-0410-8352-f15c8da7e697