From Rick Pingry, "n my application I am walking through the scene graph looking for nodes of a specific name.  I noticed there were times when some of the nodes were not named what I had expected.   Upon reviewing the code, I noticed this was happening when there was a forward reference to a parented object.
The existing code uses a map with object ID's.  The original code only applied the name to a node for a new ID only if the node did not already exist in the map.  The problem was that there was another part of the code (when a parent was forward declared) that also created the node with the ID in the map, but it did not know its name.  I simply made sure that the name was set regardless of whether or not the node was already in the map. "


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@5806 16af8721-9629-0410-8352-f15c8da7e697