From Mathias Froelich, "have done an interface to change the mouse cursor in the X11 graphics window. The win32 implementation is still in its original shape since I have no win32 implementation available.
I have chosen the enum approach for the first cut. That is benefitial since
the user does not need to track creation of mouse cursors for different
windows and displays in presence of multiple viewer windows.

The default set of available mouse shapes is the same set that was available
with glut. That set served many OpenGL applications well, so the hope is that
this is enough.

Even though, that implementation is still extensible:
I have digged out the way SDL defines new mouse cursors and added a still
documented out function prototype in the GraphicsWindow that can be used to
extend the current implemtation for arbitrary mouse shapes. That is not
implemented yet.

I hope that somebody with a win32 test system can catch up that implementation
on win32."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@6804 16af8721-9629-0410-8352-f15c8da7e697