From Dan Minor, "osgText::Text does not currently have a getAxisAlignment method.  The attached code adds this, along with a member variable to keep track of the setting.  It is based on the latest subversion version, and was tested by creating a new text object with the same axis alignment as an existing one (e.g. new_text->setAxisAlignment(old_text->getAxisAlignment()); )."
From Robert Osfield, " I originally didn't add a getAxisAlignment()
as all setAxisAlignment does is set the Rotation member variable, and
potentially one could apply user defined Rotation setting after the
setAxisAlignment() which would bring it out of sync with the
setAxisAlignment.

Rather than reject your submission on the ground of potentially
getting out of sync and therefore misleading users I've added a
USED_DEFINED_ROTATION to AxisAlignment enum, and set this in the
serRotation and then override this setting of _axisAlignment in the
setAxisAlingment method.  I've also removed the lazy updating
optimization you've added to the top of setAxisAlignment to avoid
potential problems as well."



git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@7508 16af8721-9629-0410-8352-f15c8da7e697