From Gino van den Bergen, "I've added the following features to the OSG VRML plugin: * Support for Box, Sphere, Cone, and Cylinder. These nodes are converted to osg::Geometry conform the VRML97 spec. * Backface culling is enabled/disabled according to the "solid" flag for geometries that are converted from IndexFaceSets. * PROTO instances can now be used for "appearance" and "geometry" fields in a Shape node.
The file ReaderWriterVRML2.cpp is adapted for the latest stable public
release:

http://www.openscenegraph.org/svn/osg/OpenSceneGraph/tags/OpenSceneGraph-2.2.0

The changes where needed for being able to read VRML file which are
output by VMD (http://www.ks.uiuc.edu/Research/vmd/). A sample VRML file
is enclosed in this submission.

The plugin has been tested against a number of VRML samples that include
texturing. The texturing is found to be VRML97 compliant for all added
geometry nodes.
"


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@7638 16af8721-9629-0410-8352-f15c8da7e697