From Jason Daly, "Somehow, one of the shaders in the bsp plugin started behaving incorrectly between the original submission and now.  I suspect it has to do with a typo in the setup of the shader that was fixed at some point, or it happened during the recent warnings purge.  In any case, I had to invert the usage of a parameter in the shader to make it behave properly again.
The vtf plugin wasn't working in Windows due to OS differences in the byte-packing of the header structure (on Windows, the big block read was causing a buffer overrun).  I fixed this by reading the structure from the file field by field.  It's now happy on both Linux and Windows."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@9586 16af8721-9629-0410-8352-f15c8da7e697